+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log  
 
+## [2017-11-10]  
++ **Added**  
+  - `Dockerfile` and `bootstrap` for making a docker container for running codes  
+
 ## [2017-11-08]  
 + **Added**  
   - `The JavaScript Console` in chapter 01